Se inicia implementacion de manejador de CSVs. Se modifica valores bottom y top de intervalos de int32 a float64 para aceptar valores de Infinidad. Se realiza metodo de parsing para manejar Infinity, ya que este valor no esta soportado por los metodos de parsing de JSON float64 de golang.

+++ Backend/Data/tariff_mileage.go
@@ -18,7 +18,7 @@ package data
 type TariffMileage struct {
 	Id          int64
 	VehicleType string
-	Bottom      int32
-	Top         int32
+	Bottom      float64
+	Top         float64
 	Value       float64
 }
